+++ packages/components/components/elvis-dropdown/src/react/styledComponents.tsx
@@ -6,6 +6,7 @@ const colors = {
   elviaCharge: getColor('elvia-charge'),
   elviaWhite: getColor('white'),
   elviaBlack: getColor('black'),
+  elviaFocusOutline: getColor('focus-outline'),
 };
 
 const typography = {
@@ -26,7 +27,7 @@ export const DropdownWrapper = styled.span`
     .ewc-dropdown__control {
       border: 2px solid ${colors.elviaCharge};
       padding: 0px;
-      outline: 2px solid #0064fa;
+      outline: 2px solid ${colors.elviaFocusOutline};
       outline-offset: 2px;
     }
   }
@@ -49,42 +50,13 @@ export const DropdownLabel = styled.label`
   text-align: left;
 `;
 
-type DropdownIndicatorIconProps = {
-  isDisabled: boolean;
-  isCompact: boolean;
-  menuIsOpen: boolean;
-};
-
-export const DropdownIndicatorIcon = styled.i<DropdownIndicatorIconProps>`
-  background-image: ${(props: { isDisabled: boolean }) =>
-    !props.isDisabled
-      ? `url("data:image/svg+xml,%3csvg width='24' height='24' fill='none' xmlns='http://www.w3.org/2000/svg'%3e%3cpath fill-rule='evenodd' clip-rule='evenodd' d='M.389 5.869a1.328 1.328 0 011.878 0L12 15.6l9.733-9.732a1.328 1.328 0 011.878 1.878L13.443 17.915h-.001a2.04 2.04 0 01-2.885 0L.39 7.747a1.328 1.328 0 010-1.878z' fill='black'/%3e%3c/svg%3e")`
-      : `url("data:image/svg+xml,%3csvg width='24' height='24' fill='%23BDBDBD' xmlns='http://www.w3.org/2000/svg'%3e%3cpath fill-rule='evenodd' clip-rule='evenodd' d='M.389 5.869a1.328 1.328 0 011.878 0L12 15.6l9.733-9.732a1.328 1.328 0 011.878 1.878L13.443 17.915h-.001a2.04 2.04 0 01-2.885 0L.39 7.747a1.328 1.328 0 010-1.878z' fill='%23BDBDBD'/%3e%3c/svg%3e")`};
-  height: ${(props: { isCompact: boolean }) => (props.isCompact ? '16px' : '20px')};
-  width: ${(props: { isCompact: boolean }) => (props.isCompact ? '16px' : '20px')};
-  background-size: contain;
-  background-position: center;
-  background-repeat: no-repeat;
-  display: inline-block;
-  transform: ${(props: { menuIsOpen: boolean }) => (props.menuIsOpen ? 'rotate(180deg)' : 'none')};
-  transition: 'transform 250ms';
-`;
-
 export const DropdownErrorMessageWrapper = styled.div`
   display: flex;
   flex-direction: row;
   align-items: center;
   margin-top: 4px;
 `;
-export const DropdownErrorMessageIcon = styled.i`
-  background-image: url("data:image/svg+xml,%3csvg viewBox='0 0 24 24' aria-hidden='true' width='24' height='24' fill='%23EE0701' xmlns='http://www.w3.org/2000/svg'%3e%3cg clip-path='url(%23clip0)' fill='%23EE0701'%3e%3cpath d='M12 23.999c-6.617 0-12-5.383-12-12s5.383-12 12-12 12 5.383 12 12-5.383 12-12 12zm0-22.5c-5.79 0-10.5 4.71-10.5 10.5s4.71 10.5 10.5 10.5 10.5-4.71 10.5-10.5-4.71-10.5-10.5-10.5z'/%3e%3cpath d='M16.5 17.249a.743.743 0 01-.53-.22L12 13.06l-3.97 3.97a.744.744 0 01-1.06 0 .752.752 0 010-1.061l3.97-3.97-3.97-3.97a.743.743 0 01-.22-.53c0-.2.078-.389.22-.53a.743.743 0 01.53-.22c.2 0 .389.078.53.22l3.97 3.97 3.97-3.97a.744.744 0 011.06 0c.142.141.22.33.22.53s-.078.389-.22.53l-3.97 3.97 3.97 3.97a.752.752 0 010 1.061.746.746 0 01-.53.219z'/%3e%3c/g%3e%3cdefs%3e%3cclipPath id='clip0'%3e%3cpath d='M0 0h24v24H0V0z' fill='%23EE0701'/%3e%3c/clipPath%3e%3c/defs%3e%3c/svg%3e");
-  height: 16px;
-  width: 16px;
-  background-size: contain;
-  background-position: center;
-  background-repeat: no-repeat;
-  display: inline-block;
-`;
+
 export const DropdownErrorMessageText = styled.span`
   ${typography.textSm}
   font-style: normal;
@@ -181,7 +153,7 @@ export const DropdownCheckboxMark = styled.span<DropdownCheckboxMarkProps>`
   max-height: ${(props: { isCompact: boolean }) => (props.isCompact ? '16px' : '24px')};
   border-radius: 3px;
   border: 1px solid ${colors.elviaBlack};
-  background: #fff;
+  background: ${colors.elviaWhite};
   position: relative;
   ${(props: { isCompact: boolean; isSelected: boolean }) =>
     decideCheckMarkCompactAndSelectedStyle(props.isCompact, props.isSelected)}